chore: rewrite VerticalBarOldActions
