+++ apps/meteor/app/ui-message/client/actionButtons/tabbar.ts
@@ -14,7 +14,7 @@ export const onAdded = (button: IUIActionButton): void =>
 		applyButtonFilters(button, room)
 			? {
 					id: button.actionId,
-					icon: '', // Apps won't provide icons for now
+					icon: undefined, // Apps won't provide icons for now
 					order: 300, // Make sure the button only shows up inside the room toolbox
 					title: t(Utilities.getI18nKeyForApp(button.labelI18n, button.appId)) as any,
 					// Filters were applied in the applyButtonFilters function
